Refactor: DTO 데이터 타입을 엔티티 및 Enum 구조에 맞춰 정밀 조정

+++ src/main/java/org/example/knockin/dto/ModifyPreferencesAllDto.java
@@ -9,8 +9,8 @@ import java.util.List;
 public class ModifyPreferencesAllDto {
     @Data
     public static class Request {
-        private List<Integer> lifestyles;
-        private List<Integer> conditions;
+        private List<Long> lifestyles;
+        private List<Long> conditions;
     }
 
     @Data
